change model add dependency

+++ lambda_function.py
@@ -1,10 +1,11 @@
 import json
+import numpy as np
 
 def lambda_handler(event, context):
     if 'body' in event.keys():
         data = json.loads(event['body'])
         if 'input' in data.keys():
-            input_data = int(data['input'])
+            input_data = np.float(data['input'])
             prediction = 5 * input_data + 1
             return prediction
     else:
